adding files

+++ app/Services/Maps/SetUsersLocationService.php
@@ -13,8 +13,8 @@ class SetUsersLocationService extends Service
         $user_id = auth("sanctum")->user()->id;
         $updated = User::where("id",$user_id)->update([
             'address'   => $request->address,
-            'latitude'  => (double)bcdiv($request->latitude, 1, 6),
-            'longitude' => (double)bcdiv($request->longitude, 1, 6),
+            'latitude'  => bcdiv($request->latitude, 1, 6),
+            'longitude' => bcdiv($request->longitude, 1, 6),
         ]);
         if($updated) {
             return [
